Added analyses for PDO and MEI Changed data import structure
Wavelet analyses were added for PDO and MEI timeseries. The way data are imported was also changed because so many data files are now being imported. The code used to have a data import line for each file. This commit changes that so that a list of all files matching the datafile name structure is created and that list is used to import each .csv file into a dataframe.